Add files via upload
In the current digital era, plagiarism detection techniques are crucial for maintaining the 
originality and integrity of written work. These programs use a large database of previously 
published works, such as scholarly articles, websites, and publications, to compare texts for 
similarities. Plagiarism analyzers assist writers, students, and professionals in upholding ethical 
standards and avoiding the hazards of inadvertent or intentional plagiarism by pointing out 
possible instances of copied content .Using a large database of previously published works as 
a comparison, a plagiarism analyzer is a technology that finds instances of duplicated or 
unoriginal content. It is crucial for students, teachers, and professionals to avoid plagiarism and 
follow ethical standards in writing since it contributes to ensuring the integrity and originality 
of written work. Plagiarism is a serious academic and professional offense that can have severe 
consequences ,including loss of credibility ,academic penalties, and even legal repercussions. 
With the increasing availability of online resources and digital content, the risk of unintentional 
plagiarism has also risen.